Moved common.utils to package for ease of maintenance
