+++ MyCapytain/common/reference/__init__.py
@@ -8,4 +8,4 @@
 """
 from ._base import NodeId, BaseCitationSet, BaseReference, BaseReferenceSet
 from ._capitains_cts import Citation, CtsReference, CtsReferenceSet, URN
-from ._dts_1 import DtsCitation, DtsCitationSet
+from ._dts_1 import DtsCitation, DtsCitationSet, DtsReference, DtsReferenceSet
