+++ tests/resolvers/dts/api_v1/data/root.json
@@ -1,0 +1,8 @@
+{
+  "@context": "/api/dts/contexts/EntryPoint.jsonld",
+  "@id": "/api/dts/",
+  "@type": "EntryPoint",
+  "collections": "/api/dts/collections/",
+  "documents": "/api/dts/documents/",
+  "navigation" : "/api/dts/navigation"
+}
