+++ tests/resources/texts/remote/test_cts.py
@@ -9,7 +9,7 @@ from MyCapytain.resources.texts.remote.cts import CtsPassage, CtsText
 from MyCapytain.retrievers.cts5 import HttpCtsRetriever
 from MyCapytain.common.reference._capitains_cts import CtsReference, URN, Citation
 from MyCapytain.common.metadata import Metadata
-from MyCapytain.common.utils import xmlparser
+from MyCapytain.common.utils.xml import xmlparser
 from MyCapytain.common.constants import XPATH_NAMESPACES, Mimetypes, RDF_NAMESPACES
 from MyCapytain.errors import MissingAttribute
 import mock
