+++ tests/retrievers/test_dts.py
@@ -3,9 +3,8 @@ import unittest
 import responses
 
 from MyCapytain.retrievers.dts import HttpDtsRetriever
-from MyCapytain.common.utils import _Navigation
+from MyCapytain.common.utils._http import _Navigation, parse_pagination
 from urllib.parse import parse_qs, urlparse, urljoin
-from MyCapytain.common.utils import parse_pagination
 
 _SERVER_URI = "http://domainname.com/api/dts/"
 patch_args = ("MyCapytain.retrievers.dts.requests.get", )
